adding input options

+++ stdafx.h
@@ -10,14 +10,15 @@
 #include <forward_list>
 #include <vector>
 #include <deque>
+#include <climits>
 
 using namespace std;
 
 typedef unsigned int data_t;
 typedef data_t** dataset_t;
 
-typedef unsigned int row_t;
-typedef unsigned int col_t;
+typedef size_t row_t;
+typedef size_t col_t;
 
 struct item_t
 {
